Rename src/rpclib/aux to src/rpclib/auxm Windows does not allow files named aux. So rename aux to auxm. Ref: http://support.microsoft.com/kb/74496
