+++ src/rpclib/server/twisted/__init__.py
@@ -35,7 +35,7 @@ from twisted.web.iweb import UNKNOWN_LENGTH
 from zope.interface import implements
 
 
-from rpclib import aux
+from rpclib import auxm
 from rpclib.server.http import HttpMethodContext
 from rpclib.server.http import HttpBase
 
@@ -156,7 +156,7 @@ class TwistedWebResource(Resource):
         p_ctx.out_object = error
         self.http_transport.get_out_string(p_ctx)
 
-        aux.process_contexts(self.http_transport, others, error=error)
+        auxm.process_contexts(self.http_transport, others, error=error)
 
         return ''.join(p_ctx.out_string)
 
@@ -182,7 +182,7 @@ class TwistedWebResource(Resource):
 
         self.http_transport.get_out_string(p_ctx)
 
-        aux.process_contexts(self.http_transport, others)
+        auxm.process_contexts(self.http_transport, others)
 
         def _cb_request_finished(request):
             request.finish()
